// types/tools/new-tools-dto.ts
export type NewToolsDto = {
  tool_id?: number;
  tool_name: string;
  tool_plan: string;
  tool_mode: string;
  tool_url: string;
  tool_auth: string;
  tool_login_url: string;
  tool_image: string;
  tool_day_price: number;
  tool_month_price: number;
  tool_year_price: number;
  tool_none_price_month: number;
  tool_none_price_year: number;
  tool_data: string;
  tool_local_storage: string;
  tool_blocked_urls: string;
  tool_blocked_elements: string;
  tool_external_links: string;
  tool_content: string;
  tool_description: string;
  isActive?: boolean;
  isLink?: boolean;
  cookie_1?: string;
  cookie_2?: string;
  cookie_3?: string;
  cookie_4?: string;
  cookie_5?: string;
  cookie_6?: string;
  cookie_7?: string;
  cookie_8?: string;
  cookie_9?: string;
  cookie_10?: string;
  cookie_11?: string;
  cookie_12?: string;
  cookie_13?: string;
  cookie_14?: string;
  cookie_15?: string;
  cookie_16?: string;
  cookie_17?: string;
  cookie_18?: string;
  cookie_19?: string;
  cookie_20?: string;

  local_storage_1?: string;

  blocked_url_1?: string;
  blocked_url_2?: string;
  blocked_url_3?: string;
  blocked_url_4?: string;
  blocked_url_5?: string;
  blocked_url_6?: string;
  blocked_url_7?: string;
  blocked_url_8?: string;
  blocked_url_9?: string;
  blocked_url_10?: string;

  blocked_element_1?: string;
  blocked_element_2?: string;
  blocked_element_3?: string;
  blocked_element_4?: string;
  blocked_element_5?: string;
  blocked_element_6?: string;
  blocked_element_7?: string;
  blocked_element_8?: string;
  blocked_element_9?: string;
  blocked_element_10?: string;

  external_link_1?: string;
  external_link_2?: string;
  external_link_3?: string;
  external_link_4?: string;
  external_link_5?: string;
  external_link_6?: string;
  external_link_7?: string;
  external_link_8?: string;
  external_link_9?: string;
  external_link_10?: string;
};